requirements.txt
amqp==2.6.1
billiard==3.6.3.0
celery==4.4.2
certifi==2020.11.8
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
coveralls==2.2.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.1.4
django-admin-ordering==0.12.0
django-admin-rangefilter==0.5.4
django-allauth==0.38.0
django-ckeditor==5.6.1
django-cors-headers==2.4.0
django-decouple==2.1
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.5
django-enumchoicefield==1.0.2
django-filter==2.0.0
django-haystack==2.8.1
django-import-export==1.1.0
django-js-asset==1.2.2
django-model-utils==3.1.2
django-nine==0.2.2
django-nose==1.4.6
django-redis==4.10.0
django-rest-auth==0.9.3
django-storages==1.8
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
docopt==0.6.2
drf-haystack==1.8.6
drf-yasg==1.17.0
elasticsearch==7.1.0
elasticsearch-dsl==7.1.0
et-xmlfile==1.0.1
factory-boy==2.11.1
Faker==0.9.1
idna==2.10
importlib-metadata==3.1.1
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
nose==1.3.7
oauthlib==3.1.0
packaging==20.7
Pillow==5.3.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.10.3
python3-openid==3.2.0
pytz==2020.4
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
tablib==2.0.0
text-unidecode==1.2
uritemplate==3.0.1
urllib3==1.26.2
vine==1.3.0
zipp==3.4.0
